requirements.txt
# MegaFlow2D==0.6.0
setuptools==69.2.0
scikit-learn==1.3.2
pandas==2.1.4
transformers==4.39.2
sentence-transformers==2.3.1
tensorboardX==2.6.1
accelerate==0.28.0
deepspeed==0.14.0
tqdm==4.66.2
pytest==7.4.3
scipy==1.12.0
seaborn==0.13.2
pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
wandb==0.16.6
bitsandbytes==0.42.0
peft==0.9.0
packaging==24.0
git+https://github.com/Maccy-Z/c_print.git
opencv-python==4.9.0.80
flash-attn==2.5.6
torchdiffeq
natsort